Clean: split cmd_struct_ops.c to 2 files and add doc
